pyproject.toml
[tool.poetry]
name = "ractf-backend"
version = "0.1.0"
description = "The Django backend for RACTF."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aioredis = "1.3.1"
asgiref = "3.2.3"
async-timeout = "3.0.1"
attrs = "19.3.0"
autobahn = "20.2.1"
Automat = "20.2.0"
certifi = "2019.11.28"
cffi = "1.14.0"
channels = "2.4.0"
channels-redis = "2.4.1"
chardet = "3.0.4"
constantly = "15.1.0"
coverage = "5.0.3"
cryptography = "2.8"
daphne = "2.5.0"
Django = "3.0.5"
django-cors-headers = "3.2.1"
django-redis = "4.11.0"
django-redis-cache = "2.1.1"
django-rest-framework = "0.1.0"
django-silk = "4.0.1"
django-storages = "1.9.1"
django-zxcvbn-password-validator = "1.3.0"
djangorestframework = "3.11.0"
expiringdict = "1.2.0"
gprof2dot = "2019.11.30"
hiredis = "1.0.1"
hyperlink = "19.0.0"
idna = "2.8"
incremental = "17.5.0"
Jinja2 = "2.11.2"
MarkupSafe = "1.1.1"
more-itertools = "8.2.0"
msgpack = "0.6.2"
multidict = "4.7.4"
packaging = "20.3"
pluggy = "0.13.1"
py = "1.8.1"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.19"
Pygments = "2.6.1"
PyHamcrest = "2.0.0"
pyOpenSSL = "19.1.0"
pyotp = "2.3.0"
pyparsing = "2.4.6"
python-dateutil = "2.8.1"
pytz = "2019.3"
redis = "3.5.0"
requests = "2.22.0"
requests-unixsocket = "0.2.0"
serpy = "0.3.1"
service-identity = "18.1.0"
six = "1.14.0"
sqlparse = "0.3.0"
Twisted = "19.10.0"
txaio = "20.1.1"
typing = "3.7.4.1"
urllib3 = "1.25.8"
wcwidth = "0.1.8"
yarl = "1.4.2"
"zope.interface" = "4.7.1"
zxcvbn = "4.4.28"
sentry-sdk = "^0.16.3"
gunicorn = "^20.0.4"
boto3 = "^1.14.33"
psycopg2-binary = "^2.8.5"
django-filter = "^2.3.0"

[tool.poetry.dev-dependencies]

[tool.black]
exclude = 'migrations'

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
